books_database/requirements.txt
grpcio==1.78.0
grpcio-tools==1.78.0
watchdog==6.0.0
